requirements.txt
biopython==1.79
-f https://download.pytorch.org/whl/torch_stable.html
pytorch-lightning==1.6.4
scikit-learn==1.1.1
scipy==1.8.1
tqdm==4.64.0
pandas==1.4.2
numpy==1.22.4
setuptools==59.5.0
easydict==1.10
